# requestly/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "JobPilot API",
    "version": "1.0.0"
  },
  "paths": {
    "/api/job": {
      "get": {
        "tags": [
          "job"
        ],
        "summary": "List Jobs",
        "operationId": "list_jobs_api_job_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/JobStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "shortlisted",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Shortlisted"
            }
          },
          {
            "name": "company",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Company"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ]
      }
    },
    "/api/job/createJob": {
      "post": {
        "tags": [
          "job"
        ],
        "summary": "Create Job",
        "operationId": "create_job_api_job_createJob_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobCreate"
              }
            }
          }
        }
      }
    },
    "/api/job/getJobDescription/{job_id}": {
      "get": {
        "tags": [
          "job"
        ],
        "summary": "Get Job Description",
        "description": "What the matching step reads. Same row as `getJob`, with the JD prose.",
        "operationId": "get_job_description_api_job_getJobDescription__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PydanticObjectId"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ]
      }
    },
    "/api/job/getJob/{job_id}": {
      "get": {
        "tags": [
          "job"
        ],
        "summary": "Get Job",
        "operationId": "get_job_api_job_getJob__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PydanticObjectId"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ]
      }
    },
    "/api/job/updateJob/{job_id}": {
      "patch": {
        "tags": [
          "job"
        ],
        "summary": "Update Job",
        "operationId": "update_job_api_job_updateJob__job_id__patch",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PydanticObjectId"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobUpdate"
              }
            }
          }
        }
      }
    },
...
    "/api/application/applicant": {
      "get": {
        "tags": [
          "application"
        ],
        "summary": "Get Applicant",
        "operationId": "get_applicant_api_application_applicant_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ]
      },
      "put": {
        "tags": [
          "application"
        ],
        "summary": "Save Applicant",
        "operationId": "save_applicant_api_application_applicant_put",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplicantFields"
              }
            }
          }
        }
      }
    },
    "/api/application/answer-bank": {
      "get": {
        "tags": [
          "application"
        ],
        "summary": "Answer Bank",
        "description": "Read by the extension's background worker for screening questions.",
        "operationId": "answer_bank_api_application_answer_bank_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ]
      }
    },
    "/api/application/for-job/{job_id}": {
      "get": {
        "tags": [
          "application"
        ],
        "summary": "Get For Job",
        "operationId": "get_for_job_api_application_for_job__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PydanticObjectId"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ]
      }
    },
    "/api/application/fill/{application_id}": {
      "post": {
        "tags": [
          "application"
        ],
        "summary": "Record Fill",
        "operationId": "record_fill_api_application_fill__application_id__post",
        "parameters": [
          {
            "name": "application_id",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PydanticObjectId"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplicationFill"
              }
            }
          }
        }
      }
    },
    "/api/application/status/{application_id}": {
      "patch": {
        "tags": [
          "application"
        ],
        "summary": "Set Status",
        "operationId": "set_status_api_application_status__application_id__patch",
        "parameters": [
          {
            "name": "application_id",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PydanticObjectId"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StatusTransition"
              }
            }
          }
        }
      }
    },
    "/api/application": {
      "get": {
        "tags": [
          "application"
        ],
        "summary": "List Applications",
        "operationId": "list_applications_api_application_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ApplicationStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ]
      }
    },
    "/api/application/stageApplication": {
      "post": {
        "tags": [
          "application"
        ],
        "summary": "Stage Application",
        "operationId": "stage_application_api_application_stageApplication_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplicationStage"
              }
            }
          }
        }
      }
    },
    "/api/application/getApplication/{application_id}": {
      "get": {
        "tags": [
          "application"
        ],
        "summary": "Get Application",
        "operationId": "get_application_api_application_getApplication__application_id__get",
        "parameters": [
          {
            "name": "application_id",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PydanticObjectId"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ]
      }
    },
    "/api/event": {
      "get": {
        "tags": [
          "event"
        ],
        "summary": "List Events",
        "operationId": "list_events_api_event_get",
        "parameters": [
          {
            "name": "jobId",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/PydanticObjectId"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Jobid"
            }
          },
          {
            "name": "eventType",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/EventType"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Eventtype"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ]
      }
    },
    "/api/event/appendEvent": {
      "post": {
        "tags": [
          "event"
        ],
        "summary": "Append Event",
        "operationId": "append_event_api_event_appendEvent_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventAppend"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApplicantFields": {
        "properties": {
          "address": {
            "$ref": "#/components/schemas/Address"
          },
          "demographics": {
            "$ref": "#/components/schemas/Demographics"
          },
          "totalExperienceYears": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 70.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Totalexperienceyears"
          },
          "noticePeriod": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Noticeperiod"
          },
          "earliestStartDate": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Earlieststartdate"
          },
          "currentSalary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Currentsalary"
          },
          "expectedSalary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expectedsalary"
          },
          "workAuthorization": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workauthorization"
          },
          "requiresSponsorship": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requiressponsorship"
          },
          "willingToRelocate": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Willingtorelocate"
          },
          "howHeard": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Howheard"
          },
          "referredBy": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Referredby"
          }
        },
        "type": "object",
        "title": "ApplicantFields",
        "description": "The facts every application form asks for and no resume carries, so the\nprofile has nowhere to put them. Stored once, answered everywhere.\n\nThis is not `P4-02`: those are discovery preferences (which jobs to look for).\nThese are answers about you, used only when filling a form."
      },
      "ApplicationFill": {
        "properties": {
          "fieldsFilled": {
            "items": {
              "$ref": "#/components/schemas/FieldFill"
            },
            "type": "array",
            "title": "Fieldsfilled"
          },
          "screeningAnswers": {
            "items": {
              "$ref": "#/components/schemas/ScreeningAnswer"
            },
            "type": "array",
            "title": "Screeninganswers"
          }
        },
        "type": "object",
        "title": "ApplicationFill",
        "description": "The extension reporting what it filled. Reporting only \u2014 it submits nothing."
      },
      "ApplicationStage": {
        "properties": {
          "jobId": {
            "$ref": "#/components/schemas/PydanticObjectId"
          },
          "resumeId": {
            "$ref": "#/components/schemas/PydanticObjectId"
          },
          "texPath": {
            "type": "string",
            "title": "Texpath"
          },
          "ats": {
            "$ref": "#/components/schemas/AtsPlatform",
            "default": "other"
          },
          "applyUrl": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2083,
                "minLength": 1,
                "format": "uri"
              },
              {
                "type": "null"
              }
            ],
            "title": "Applyurl"
          }
        },
        "type": "object",
        "required": [
          "jobId",
          "resumeId",
          "texPath"
        ],
        "title": "ApplicationStage",
        "description": "What `resume` posts once a .tex exists for the job."
      },
      "ApplicationStatus": {
        "type": "string",
        "enum": [
          "staged",
          "applied",
          "viewed",
          "interview",
          "offer",
          "rejected",
          "withdrawn"
        ],
        "title": "ApplicationStatus",
        "description": "STAGED is pre-submit. Everything after it is a human's doing."
      },
...
      "EventAppend": {
        "properties": {
          "eventType": {
            "$ref": "#/components/schemas/EventType"
          },
          "actor": {
            "$ref": "#/components/schemas/Actor"
          },
          "jobId": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PydanticObjectId"
              },
              {
                "type": "null"
              }
            ]
          },
          "applicationId": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PydanticObjectId"
              },
              {
                "type": "null"
              }
            ]
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          }
        },
        "type": "object",
        "required": [
          "eventType",
          "actor"
        ],
        "title": "EventAppend",
        "description": "Append-only: there is no update or delete on an event."
      },
      "EventType": {
        "type": "string",
        "enum": [
          "job_discovered",
          "job_shortlisted",
          "match_scored",
          "keywords_selected",
          "keywords_skipped",
          "resume_tailored",
          "application_staged",
          "application_filled",
          "status_changed",
          "follow_up_drafted",
          "follow_up_sent",
          "sheet_synced"
        ],
        "title": "EventType"
      },
...
      "JobCreate": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "company": {
            "$ref": "#/components/schemas/Company"
          },
          "location": {
            "type": "string",
            "title": "Location"
          },
          "jobType": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JobType"
              },
              {
                "type": "null"
              }
            ]
          },
          "workMode": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WorkMode"
              },
              {
                "type": "null"
              }
            ]
          },
          "experienceBand": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Experienceband"
          },
          "salaryText": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Salarytext"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "responsibilities": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Responsibilities"
          },
          "requirements": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Requirements"
          },
          "source": {
            "$ref": "#/components/schemas/JobSource"
          },
          "sourceUrl": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2083,
                "minLength": 1,
                "format": "uri"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sourceurl"
          },
          "postedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Postedat"
          },
          "deadlineAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deadlineat"
          },
          "applicantCount": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Applicantcount"
          },
          "jdText": {
            "type": "string",
            "title": "Jdtext"
          },
          "dedupHash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deduphash"
          }
        },
        "type": "object",
        "required": [
          "title",
          "company",
          "location",
          "source",
          "jdText"
        ],
        "title": "JobCreate",
        "description": "The posting as discovery found it."
      },
...
      "JobStatus": {
        "type": "string",
        "enum": [
          "new",
          "reviewed",
          "tailored",
          "applied",
          "archived"
        ],
        "title": "JobStatus",
        "description": "Pipeline column on the dashboard board."
      },
      "JobType": {
        "type": "string",
        "enum": [
          "full_time",
          "contract",
          "part_time",
          "internship"
        ],
        "title": "JobType"
      },
      "JobUpdate": {
        "properties": {
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JobStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "shortlisted": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shortlisted"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "responsibilities": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Responsibilities"
          },
          "requirements": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requirements"
          }
        },
        "type": "object",
        "title": "JobUpdate",
        "description": "Every field optional: a PATCH sends only what changed."
      },
...
      "PydanticObjectId": {
        "type": "string",
        "maxLength": 24,
        "minLength": 24,
        "pattern": "^[0-9a-f]{24}$",
        "example": "5eb7cf5a86d9755df3a6c593"
      },
...
      "StatusTransition": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/ApplicationStatus"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "confirmedByUser": {
            "type": "boolean",
            "title": "Confirmedbyuser",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "StatusTransition"
      },
...
    }
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8000",
      "description": "local \u2014 server/main.py"
    }
  ]
}